tests+cleanup: co-locate next tests, single impl per type, lighter comments
- Move the repeat-last successor tests into the existing eq.rs test module
  instead of a second free-standing test module.
- Merge the duplicate per-type impl blocks so each type has one impl block
  next to its definition (EqMaybePacked and SplitEq in split_eq, SvoPoint
  in svo).
- Add small co-located unit tests for the new entry points: Point::eval_next
  (empty and single-variable known answers), Poly::new_next_from_point
  (shift-and-repeat table), and Poly::eval_next_base (matches the explicit
  shifted-table evaluation).
- Trim the eval_select and eval_next inline comments to one crisp idea per
  line, dropping the multi-line prose blocks.

multilinear-util 141 and sumcheck 201 tests pass; fmt and clippy clean.
