+++ multilinear-util/src/split_eq/mod.rs
@@ -448,9 +448,7 @@ impl<F: Field, EF: ExtensionField<F>> SplitEq<F, EF> {
                 });
         }
     }
-}
 
-impl<F: Field, EF: ExtensionField<F>> SplitEq<F, EF> {
     /// Evaluates a base-field polynomial against the repeat-last successor weights.
     ///
     /// Each hypercube row is read at its successor, with the maximal row repeating itself:
@@ -950,364 +948,6 @@ mod tests {
         // Each element should equal its base-field counterpart promoted to the extension.
         for (c, &p) in compressed.iter().zip(poly.iter()) {
             assert_eq!(*c, EF::from(p));
-        }
-    }
-}
-
-#[cfg(test)]
-mod next_tests {
-    use alloc::vec::Vec;
-
-    use itertools::Itertools;
-    use p3_baby_bear::BabyBear;
-    use p3_field::extension::BinomialExtensionField;
-    use p3_field::{ExtensionField, Field, PrimeCharacteristicRing, dot_product};
-    use proptest::prelude::*;
-    use rand::rngs::SmallRng;
-    use rand::{RngExt, SeedableRng};
-
-    use crate::point::Point;
-    use crate::poly::Poly;
-    use crate::split_eq::SplitEq;
-    type F = BabyBear;
-    type EF = BinomialExtensionField<F, 4>;
-
-    // Plain dense evaluation: dot the equality table for the point with the evals.
-    fn eval_reference<F: Field, EF: ExtensionField<F>>(evals: &[F], point: &[EF]) -> EF {
-        let eq = Poly::new_from_point(point, EF::ONE);
-        dot_product(eq.iter().copied(), evals.iter().copied())
-    }
-
-    // Dense successor evaluation built the slow but obvious way.
-    fn eval_next_reference<F: Field, EF: ExtensionField<F>>(evals: &[F], point: &[EF]) -> EF {
-        let mut shifted = evals.to_vec();
-        // Preserve the maximal row so it can repeat instead of wrapping around.
-        let last = *shifted.last().unwrap();
-        // Rotate every row down by one so row x now holds the old row x + 1.
-        shifted.rotate_left(1);
-        // Overwrite the wrapped-in row 0 with the repeated maximal row.
-        *shifted.last_mut().unwrap() = last;
-        // Evaluate the rebuilt table the plain way.
-        eval_reference(&shifted, point)
-    }
-
-    // Dense reference for fixing the prefix variables with predecessor weights.
-    fn compress_prefix_shifted_poly_reference<F: Field, EF: ExtensionField<F>>(
-        poly: &Poly<F>,
-        point: &Point<EF>,
-    ) -> Poly<EF> {
-        assert!(point.num_variables() <= poly.num_variables());
-        // Inner variables remain free after fixing the prefix.
-        let inner_vars = poly.num_variables() - point.num_variables();
-        let inner_rows = 1 << inner_vars;
-        let split_rows = 1 << point.num_variables();
-        let eq = Poly::new_from_point(point.as_slice(), EF::ONE);
-        let mut out = Poly::<EF>::zero(inner_vars);
-
-        // Skip prefix row 0: it has no predecessor, so its shifted weight is zero.
-        for split_idx in 1..split_rows {
-            // Predecessor weight: prefix row split_idx reads equality entry split_idx - 1.
-            let weight = eq.as_slice()[split_idx - 1];
-            // The inner-variable block belonging to this prefix row.
-            let chunk = &poly.as_slice()[split_idx * inner_rows..(split_idx + 1) * inner_rows];
-            out.as_mut_slice()
-                .iter_mut()
-                .zip_eq(chunk.iter())
-                .for_each(|(out, &value)| *out += weight * value);
-        }
-
-        out
-    }
-
-    // Dense reference for fixing the suffix variables with predecessor weights.
-    fn compress_suffix_shifted_reference<F: Field, EF: ExtensionField<F>>(
-        poly: &Poly<F>,
-        point: &Point<EF>,
-    ) -> Poly<EF> {
-        assert!(point.num_variables() <= poly.num_variables());
-        let suffix_vars = point.num_variables();
-        let prefix_rows = poly.num_evals() >> suffix_vars;
-        let suffix_rows = 1 << suffix_vars;
-        let eq = Poly::new_from_point(point.as_slice(), EF::ONE);
-        let mut out = Poly::<EF>::zero(poly.num_variables() - suffix_vars);
-
-        for prefix_idx in 0..prefix_rows {
-            // Skip suffix row 0: it has no predecessor, so its shifted weight is zero.
-            for suffix_idx in 1..suffix_rows {
-                // Predecessor weight: suffix row suffix_idx reads equality entry suffix_idx - 1.
-                let weight = eq.as_slice()[suffix_idx - 1];
-                // Flat index: prefix bits sit above the suffix bits.
-                let value = poly.as_slice()[(prefix_idx << suffix_vars) | suffix_idx];
-                out.as_mut_slice()[prefix_idx] += weight * value;
-            }
-        }
-
-        out
-    }
-
-    // Dense reference for one outer chunk of the successor accumulation.
-    fn accumulate_next_chunk_reference(
-        eq: &[EF],
-        out: &mut [EF],
-        eq_weight: EF,
-        shifted_weight: EF,
-        boundary: EF,
-    ) {
-        assert_eq!(eq.len(), out.len());
-
-        for idx in 0..out.len() {
-            // Each row gets its own equality weight plus a shifted predecessor term.
-            // Row 0 has no in-chunk predecessor, so it takes the precomputed
-            // cross-chunk boundary instead.
-            out[idx] += eq_weight * eq[idx]
-                + if idx == 0 {
-                    boundary
-                } else {
-                    shifted_weight * eq[idx - 1]
-                };
-        }
-    }
-
-    // Dense reference for the interior prefix-compression accumulation.
-    fn compress_prefix_shifted_reference(eq: &[EF], out: &mut [EF], chunk: &[F], w0: EF) {
-        let size_inner = out.len();
-        // Skip the first inner block: its predecessor lives in the previous outer chunk.
-        chunk
-            .chunks(size_inner)
-            .skip(1)
-            .zip(eq.iter())
-            .for_each(|(chunk, &w1)| {
-                // Combined weight: outer prefix weight times suffix predecessor weight.
-                let w = w0 * w1;
-                out.iter_mut()
-                    .zip_eq(chunk.iter())
-                    .for_each(|(out, &value)| *out += w * value);
-            });
-    }
-
-    #[test]
-    fn test_next_closed_forms() {
-        // Invariant: four different routes to the successor opening must agree.
-        //   1. closed-form carry recurrence over the two points
-        //   2. dense successor table dotted with the equality table
-        //   3. dense successor table evaluated at the first point
-        //   4. plain equality table evaluated through the successor view
-        let mut rng = SmallRng::seed_from_u64(0);
-        // Fixture state: sweep variable counts 0..14 to cover all base-case branches.
-        for k in 0..14 {
-            let p0 = Point::<F>::rand(&mut rng, k);
-            let p1 = Point::<F>::rand(&mut rng, k);
-
-            // Route 1: closed form. Full fold yields completed plus boundary mass.
-            let (_carry, done, omega) = Point::eval_next(p1.as_slice(), p0.as_slice());
-            let e0 = done + omega;
-
-            // Route 2: materialize the successor table, then dot with the equality table.
-            let next = Poly::new_next_from_point(p1.as_slice());
-            let eq = Poly::new_from_point(p0.as_slice(), F::ONE);
-            let e1 = dot_product(eq.iter().copied(), next.iter().copied());
-            assert_eq!(e0, e1);
-
-            // Route 3: evaluate the same successor table directly at the first point.
-            let e1 = next.eval_base(&p0);
-            assert_eq!(e0, e1);
-
-            // Route 4: keep the plain equality table and apply the successor view at eval time.
-            let next = Poly::new_from_point(p0.as_slice(), F::ONE);
-            let e1 = next.eval_next_base(&p1);
-            assert_eq!(e0, e1);
-        }
-    }
-
-    #[test]
-    fn test_accumulate_next_chunk_into_matches_reference() {
-        // Invariant: both the scalar and SIMD-packed suffix tables produce the
-        // same per-chunk accumulation as the dense reference.
-        let mut rng = SmallRng::seed_from_u64(1);
-
-        // Fixture state: vary total variables 2..=10 so the suffix half spans
-        // both sub-packing-width and multi-packed-block sizes.
-        for k in 2..=10 {
-            let point = Point::<EF>::rand(&mut rng, k);
-            // Split in the middle so the suffix half has at least one variable.
-            let split_at = point.num_variables() / 2;
-            let (_prefix, suffix) = point.split_at(split_at);
-            let eq = Poly::new_from_point(suffix.as_slice(), EF::ONE);
-
-            // Random weights stand in for the prefix-driven scalars supplied at runtime.
-            let eq_weight = rng.random();
-            let shifted_weight = rng.random();
-            let boundary = rng.random();
-            // Random starting accumulator to confirm the routine adds rather than overwrites.
-            let initial = (0..eq.num_evals())
-                .map(|_| rng.random())
-                .collect::<Vec<_>>();
-
-            // Golden result from the obvious dense loop.
-            let mut expected = initial.clone();
-            accumulate_next_chunk_reference(
-                eq.as_slice(),
-                &mut expected,
-                eq_weight,
-                shifted_weight,
-                boundary,
-            );
-
-            // Scalar storage path must match the golden result.
-            let mut unpacked = initial.clone();
-            SplitEq::<F, EF>::new_unpacked(&point, EF::ONE)
-                .eq1()
-                .accumulate_next_chunk_into(
-                    unpacked.as_mut_slice(),
-                    eq_weight,
-                    shifted_weight,
-                    boundary,
-                );
-            assert_eq!(unpacked, expected);
-
-            // SIMD-packed storage path must produce identical scalars.
-            let mut packed = initial.clone();
-            SplitEq::<F, EF>::new_packed(&point, EF::ONE)
-                .eq1()
-                .accumulate_next_chunk_into(
-                    packed.as_mut_slice(),
-                    eq_weight,
-                    shifted_weight,
-                    boundary,
-                );
-            assert_eq!(packed, expected);
-        }
-    }
-
-    #[test]
-    fn test_compress_prefix_shifted_into_matches_reference() {
-        // Invariant: scalar and packed suffix tables yield the same interior
-        // prefix-compression accumulation as the dense reference.
-        let mut rng = SmallRng::seed_from_u64(2);
-
-        // Fixture state: vary total variables 2..=10 to cross packing boundaries.
-        for k in 2..=10 {
-            let point = Point::<EF>::rand(&mut rng, k);
-            // Middle split gives a non-empty suffix half to drive the weights.
-            let split_at = point.num_variables() / 2;
-            let (_prefix, suffix) = point.split_at(split_at);
-            let eq = Poly::new_from_point(suffix.as_slice(), EF::ONE);
-            // One inner block per inner variable assignment.
-            let size_inner = 1 << split_at;
-            // Random outer chunk laid out as suffix-rows blocks of inner_size each.
-            let chunk = (0..(eq.num_evals() * size_inner))
-                .map(|_| rng.random())
-                .collect::<Vec<_>>();
-            // Random outer prefix weight applied to the whole chunk.
-            let w0 = rng.random();
-            // Random starting accumulator to confirm the routine adds in place.
-            let initial = (0..size_inner).map(|_| rng.random()).collect::<Vec<_>>();
-
-            // Golden result from the obvious dense loop.
-            let mut expected = initial.clone();
-            compress_prefix_shifted_reference(eq.as_slice(), &mut expected, chunk.as_slice(), w0);
-
-            // Scalar storage path.
-            let mut unpacked = initial.clone();
-            SplitEq::<F, EF>::new_unpacked(&point, EF::ONE)
-                .eq1()
-                .compress_prefix_shifted_into(unpacked.as_mut_slice(), chunk.as_slice(), w0);
-            assert_eq!(unpacked, expected);
-
-            // SIMD-packed storage path must produce identical scalars.
-            let mut packed = initial.clone();
-            SplitEq::<F, EF>::new_packed(&point, EF::ONE)
-                .eq1()
-                .compress_prefix_shifted_into(packed.as_mut_slice(), chunk.as_slice(), w0);
-            assert_eq!(packed, expected);
-        }
-    }
-
-    proptest! {
-        #[test]
-        fn prop_eval_next_base_matches_shifted_poly_reference(k in 0usize..=14, seed in any::<u64>()) {
-            // Invariant: the split-form successor evaluation equals the dense
-            // shifted-table reference, for both scalar and packed storage.
-            let mut rng = SmallRng::seed_from_u64(seed);
-            let poly = Poly::<F>::rand(&mut rng, k);
-            let point = Point::<EF>::rand(&mut rng, k);
-            // Golden value from materializing the shifted table and evaluating it.
-            let expected = eval_next_reference(poly.as_slice(), point.as_slice());
-
-            // Scalar storage must match the golden value.
-            prop_assert_eq!(
-                expected,
-                SplitEq::<F, EF>::new_unpacked(&point, EF::ONE).eval_next_base(&poly),
-            );
-            // Packed storage must match the golden value.
-            prop_assert_eq!(
-                expected,
-                SplitEq::<F, EF>::new_packed(&point, EF::ONE).eval_next_base(&poly),
-            );
-        }
-
-        #[test]
-        fn prop_compress_prefix_shifted_matches_reference(
-            split_vars in 0usize..=8,
-            inner_vars in 0usize..=6,
-            seed in any::<u64>(),
-        ) {
-            // Invariant: prefix compression with shifted weights matches the
-            // dense reference, for both base-field and extension-field polynomials.
-            let mut rng = SmallRng::seed_from_u64(seed);
-            // Total variables = fixed prefix variables + remaining inner variables.
-            let total_vars = split_vars + inner_vars;
-            let point = Point::<EF>::rand(&mut rng, split_vars);
-
-            // Base-field input: coefficients in the base field.
-            let base_poly = Poly::<F>::rand(&mut rng, total_vars);
-            let expected = compress_prefix_shifted_poly_reference(&base_poly, &point);
-            prop_assert_eq!(
-                expected,
-                SplitEq::<F, EF>::new_packed(&point, EF::ONE)
-                    .compress_prefix_shifted(&base_poly),
-            );
-
-            // Extension-field input: exercises the all-extension storage path.
-            let ext_poly = Poly::<EF>::rand(&mut rng, total_vars);
-            let expected = compress_prefix_shifted_poly_reference(&ext_poly, &point);
-            prop_assert_eq!(
-                expected,
-                SplitEq::<EF, EF>::new_packed(&point, EF::ONE)
-                    .compress_prefix_shifted(&ext_poly),
-            );
-        }
-
-        #[test]
-        fn prop_compress_suffix_shifted_matches_reference(
-            suffix_vars in 0usize..=8,
-            prefix_vars in 0usize..=6,
-            seed in any::<u64>(),
-        ) {
-            // Invariant: suffix compression with shifted weights matches the
-            // dense reference, for both base-field and extension-field polynomials.
-            let mut rng = SmallRng::seed_from_u64(seed);
-            // Total variables = fixed suffix variables + remaining prefix variables.
-            let total_vars = suffix_vars + prefix_vars;
-            let point = Point::<EF>::rand(&mut rng, suffix_vars);
-
-            // Base-field input: coefficients in the base field.
-            let base_poly = Poly::<F>::rand(&mut rng, total_vars);
-            let expected = compress_suffix_shifted_reference(&base_poly, &point);
-            prop_assert_eq!(
-                expected,
-                SplitEq::<F, EF>::new_packed(&point, EF::ONE)
-                    .compress_suffix_shifted(&base_poly),
-            );
-
-            // Extension-field input: exercises the all-extension storage path.
-            let ext_poly = Poly::<EF>::rand(&mut rng, total_vars);
-            let expected = compress_suffix_shifted_reference(&ext_poly, &point);
-            prop_assert_eq!(
-                expected,
-                SplitEq::<EF, EF>::new_packed(&point, EF::ONE)
-                    .compress_suffix_shifted(&ext_poly),
-            );
         }
     }
 }
